Move curve_lend under curve
