+++ rotkehlchen/chain/evm/decoding/curve/lend/utils.py
