+++ rotkehlchen/tests/utils/decoders.py
@@ -25,7 +25,7 @@ def patch_decoder_reload_data(load_global_caches: list[str] | None = None) -> Ex
             new=lambda *args, **kwargs: False,
         ))
         stack.enter_context(patch(  # patch to not refresh cache by not downloading new data
-            target='rotkehlchen.chain.evm.decoding.curve_lend.decoder.should_update_protocol_cache',
+            target='rotkehlchen.chain.evm.decoding.curve.lend.decoder.should_update_protocol_cache',
             new=lambda *args, **kwargs: False,
         ))
         # patch_general and patch_unique are booleans to indicate if we want to patch
